refactor(scripts): Add TextUtils and move collection sanization
